Rename C++ Python types with ll_ prefix.
Add new Python classes inheriting from ll_ Python types and
get rid of monkey-patching.

Remove test of very old file formats that can no longer be supported.

Python wrappers to C++ simplify functions.

DiploidPopulation init keyword demesizes is gone!  (Breaks API)
