+++ fwdpy11/__init__.py
@@ -29,8 +29,6 @@ from ._demography import *  # NOQA
 from ._dev import *  # NOQA
 from ._fwdpy11 import *  # NOQA
 from ._monkeypatch import _data_matrix  # NOQA
-from ._monkeypatch import _diploid_population  # NOQA
-from ._monkeypatch import _table_collection  # NOQA
 from .discrete_demography import (  # NOQA
     DiscreteDemography,
     MassMigration,
@@ -50,7 +48,7 @@ from .genetic_map_unit import (
     BinomialPoint,
     FixedCrossovers,
 )
-from .mutation_dominance import ( # NOQA
+from .mutation_dominance import (  # NOQA
     FixedDominance,
     UniformDominance,
     ExponentialDominance,
@@ -70,9 +68,13 @@ from .genetic_values import (  # NOQA
     GBR,
     StrictAdditiveMultivariateEffects,
 )
+
+from ._types import TableCollection, DiploidPopulation  # NOQA
 from ._types.demography_debugger import DemographyDebugger  # NOQA
 from ._types.model_params import ModelParams, MutationAndRecombinationRates  # NOQA
 from ._evolvets import *  # NOQA
+
+from ._functions import simplify, simplify_tables  # NOQA
 
 if sys.version_info[0] < 3:
     raise ValueError("Python3 required!")
@@ -80,6 +82,4 @@ if sys.version_info[0] < 3:
 
 # NOTE: some operations that can be implemented efficiently
 # in Python are supplied as monkey-patches to the pybind11 classes
-_monkeypatch._diploid_population._patch_diploid_population(DiploidPopulation)  # NOQA
-_monkeypatch._table_collection._patch_table_collection(TableCollection)  # NOQA
 _monkeypatch._data_matrix._patch_data_matrix(DataMatrix)  # NOQA
