+++ fwdpy11/_functions/__init__.py
@@ -1,0 +1,1 @@
+from .simplify_tables import simplify, simplify_tables  # NOQA
