+++ fwdpy11/src/fwdpy11_types/PopulationBase.cc
@@ -34,34 +34,6 @@ struct flattened_Mutation
 
 namespace
 {
-    static const auto MCOUNTS_DOCSTRING = R"delim(
-    List of number of occurrences of elements in 
-    a population objecst "mutations" container.
-
-    The values are unsigned 32-bit integers.  
-
-    .. note::
-        Some values may be 0.  These represent *extinct* variants.  You will typically want to avoid processing such mutations.
-)delim";
-
-    static const auto FIXATIONS_DOCSTRING
-        = R"delim(A :class:`fwdpy11.VecMutation` of fixed variants.)delim";
-
-    static const auto FIXATION_TIMES_DOCSTRING
-        = R"delim(A list of fixation times corresponding to the elements in "fixations" for this type.)delim";
-
-    static const auto GAMETES_DOCSTRING
-        = R"delim(A :class:`fwdpy11.HaploidGenomeVector`.)delim";
-
-    static const auto MUTATIONS_DOCSTRING = R"delim(
-    List of :class:`fwdpy11.Mutation`.
-
-    .. note:: 
-        This list contains **both** extinct *and* extant mutations.  
-        To distinguish them, use the locations of nonzero values in "mcounts" 
-        for an instance of this type."
-    )delim";
-
     py::array
     make_flattened_Mutation_array(
         const fwdpy11::Population::mutation_container& mutations)
@@ -70,8 +42,8 @@ namespace
         vfm.reserve(mutations.size());
         for (auto&& m : mutations)
             {
-                vfm.push_back(flattened_Mutation{ m.pos, m.s, m.h, m.g, m.xtra,
-                                                  m.neutral });
+                vfm.push_back(
+                    flattened_Mutation{m.pos, m.s, m.h, m.g, m.xtra, m.neutral});
             }
         return fwdpy11::make_1d_array_with_capsule(std::move(vfm));
     }
@@ -85,44 +57,26 @@ init_PopulationBase(py::module& m)
 {
     PYBIND11_NUMPY_DTYPE(flattened_Mutation, pos, s, h, g, label, neutral);
 
-    py::class_<fwdpy11::Population>(m, "PopulationBase",
-                                    "Abstract base class for populations "
-                                    "based on :class:`fwdpy11.Mutation`")
-        .def_readonly("N", &fwdpy11::Population::N, "Current population size.")
-        .def_readonly("generation", &fwdpy11::Population::generation,
-                      "Curent generation.")
-        .def_readonly("mutations", &fwdpy11::Population::mutations,
-                      MUTATIONS_DOCSTRING)
+    py::class_<fwdpy11::Population>(m, "PopulationBase")
+        .def_readonly("_N", &fwdpy11::Population::N)
+        .def_readonly("_generation", &fwdpy11::Population::generation)
+        .def_readonly("_mutations", &fwdpy11::Population::mutations)
+        .def_property_readonly("_mutations_ndarray",
+                               [](const fwdpy11::Population& self) {
+                                   return make_flattened_Mutation_array(self.mutations);
+                               })
+        .def_property_readonly("_mcounts",
+                               [](const fwdpy11::Population& self) {
+                                   return fwdpy11::make_1d_ndarray_readonly(
+                                       self.mcounts);
+                               })
+        .def_property_readonly("_mcounts_ancient_samples",
+                               [](const fwdpy11::Population& self) {
+                                   return fwdpy11::make_1d_ndarray_readonly(
+                                       self.mcounts_from_preserved_nodes);
+                               })
         .def_property_readonly(
-            "mutations_ndarray",
-            [](const fwdpy11::Population& self) {
-                return make_flattened_Mutation_array(self.mutations);
-            },
-            R"delim(
-                               Return readonly numpy.ndarray of mutation data.
-                               The data are returned as a copy.
-
-                               The esizes and heffects fields are not part of the 
-                               array, as their size is not constant and therefore
-                               they canot be part of a numpy "dtype".
-
-                               .. versionadded:: 0.4.0
-                               )delim")
-        .def_property_readonly(
-            "mcounts",
-            [](const fwdpy11::Population& self) {
-                return fwdpy11::make_1d_ndarray_readonly(self.mcounts);
-            },
-            MCOUNTS_DOCSTRING)
-        .def_property_readonly(
-            "mcounts_ancient_samples",
-            [](const fwdpy11::Population& self) {
-                return fwdpy11::make_1d_ndarray_readonly(
-                    self.mcounts_from_preserved_nodes);
-            },
-            "The contribution that ancient samples make to mutation counts")
-        .def_property_readonly(
-            "mut_lookup",
+            "_mut_lookup",
             [](const fwdpy11::Population& pop) -> py::object {
                 py::dict rv;
                 for (std::size_t i = 0; i < pop.mutations.size(); ++i)
@@ -131,8 +85,7 @@ init_PopulationBase(py::module& m)
                             || (!pop.mcounts_from_preserved_nodes.empty()
                                 && pop.mcounts_from_preserved_nodes[i] > 0))
                             {
-                                auto pos_handle
-                                    = py::cast(pop.mutations[i].pos);
+                                auto pos_handle = py::cast(pop.mutations[i].pos);
                                 if (rv.contains(pos_handle))
                                     {
                                         py::list l = rv[pos_handle];
@@ -151,20 +104,10 @@ init_PopulationBase(py::module& m)
                         return py::none();
                     }
                 return py::object(std::move(rv));
-            },
-            R"delim(
-            Mutation position lookup table.
-            The format is a dict whose keys 
-            are mutation positions and values
-            are lists of indexes.  If no
-            extant mutations are present in the 
-            population, the value of this property is
-            None.
-            )delim")
+            })
         .def(
-            "mutation_indexes",
-            [](const fwdpy11::Population& pop,
-               const double pos) -> py::object {
+            "_mutation_indexes",
+            [](const fwdpy11::Population& pop, const double pos) -> py::object {
                 auto r = pop.mut_lookup.equal_range(pos);
                 if (r.first == r.second)
                     {
@@ -178,91 +121,32 @@ init_PopulationBase(py::module& m)
                 std::sort(begin(rv), end(rv));
                 return fwdpy11::make_1d_array_with_capsule(std::move(rv));
             },
-            R"delim(
-             Get indexes associated with a mutation position.
-             
-             :param pos: A position
-             :type pos: float
-             :return: Indexes in mutation/mutation counts container associated with pos.
-             :rtype: numpy.array
-             
-             Returns None if pos does not refer to an extant variant.  Otherwise, 
-             returns a list.
-             )delim",
             py::arg("pos"))
-        .def_readonly("haploid_genomes", &fwdpy11::Population::haploid_genomes,
-                      GAMETES_DOCSTRING)
-        .def_readonly("fixations", &fwdpy11::Population::fixations,
-                      FIXATIONS_DOCSTRING)
-        .def_readonly("fixation_times", &fwdpy11::Population::fixation_times,
-                      FIXATION_TIMES_DOCSTRING)
-        .def(
-            "find_mutation_by_key",
-            [](const fwdpy11::Population& pop,
-               const std::tuple<double, double, fwdpp::uint_t>& key,
-               const std::int64_t offset) {
-                return pop.find_mutation_by_key(key, offset);
-            },
-            py::arg("pop"), py::arg("offset") = 0,
-            R"delim(
-             Find a mutation by key.
-             
-             :param key: A mutation key. See :func:`fwdpy11.Mutation.key`.
-             :type key: tuple
-             :param offset: Offset to start search in mutation container.
-             :type offset: int
-
-             :rtype: int
-
-             :returns: Index of mutation if found, or -1 otherwise.
-
-             .. versionadded:: 0.2.0
-             )delim")
-        .def(
-            "find_fixation_by_key",
-            [](const fwdpy11::Population& pop,
-               const std::tuple<double, double, fwdpp::uint_t>& key,
-               const std::int64_t offset) {
-                return pop.find_fixation_by_key(key, offset);
-            },
-            py::arg("pop"), py::arg("offset") = 0,
-            R"delim(
-             Find a fixation by key.
-             
-             :param key: A mutation key. See :func:`fwdpy11.Mutation.key`.
-             :type key: tuple
-             :param offset: Offset to start search in fixation container.
-             :type offset: int
-
-             :rtype: int
-
-             :returns: Index of fixation if found, or -1 otherwise.
-
-             .. versionadded:: 0.2.0
-             )delim")
+        .def_readonly("_haploid_genomes", &fwdpy11::Population::haploid_genomes)
+        .def_readonly("_fixations", &fwdpy11::Population::fixations)
+        .def_readonly("_fixation_times", &fwdpy11::Population::fixation_times)
+        .def("_find_mutation_by_key",
+             [](const fwdpy11::Population& pop,
+                const std::tuple<double, double, fwdpp::uint_t>& key,
+                const std::int64_t offset) {
+                 return pop.find_mutation_by_key(key, offset);
+             })
+        .def("_find_fixation_by_key",
+             [](const fwdpy11::Population& pop,
+                const std::tuple<double, double, fwdpp::uint_t>& key,
+                const std::int64_t offset) {
+                 return pop.find_fixation_by_key(key, offset);
+             })
         // TODO: why does readwrite fail?
-        .def_readonly("tables", &fwdpy11::Population::tables,
-                      R"delim(
-                Give access to the population's 
-                :class:`fwdpy11.TableCollection`
-                )delim")
+        .def_readonly("_tables", &fwdpy11::Population::tables)
+        .def_property_readonly("_genetic_values",
+                               [](const fwdpy11::Population& self) {
+                                   return fwdpy11::make_2d_ndarray_readonly(
+                                       self.genetic_value_matrix, self.N,
+                                       self.genetic_value_matrix.size() / self.N);
+                               })
         .def_property_readonly(
-            "genetic_values",
-            [](const fwdpy11::Population& self) {
-                return fwdpy11::make_2d_ndarray_readonly(
-                    self.genetic_value_matrix, self.N,
-                    self.genetic_value_matrix.size() / self.N);
-            },
-            R"delim(
-        Return the genetic values as a readonly 2d numpy.ndarray.
-        
-        Rows are individuals.  Columns are genetic values.
-        
-        ..  versionadded 0.3.0
-        )delim")
-        .def_property_readonly(
-            "ancient_sample_genetic_values",
-            [](const fwdpy11::Population& self) {
+            "_ancient_sample_genetic_values", [](const fwdpy11::Population& self) {
                 if (self.ancient_sample_metadata_size() == 0
                     || self.ancient_sample_genetic_value_matrix.empty())
                     {
@@ -274,13 +158,5 @@ init_PopulationBase(py::module& m)
                     self.ancient_sample_metadata_size(),
                     self.ancient_sample_genetic_value_matrix.size()
                         / self.ancient_sample_metadata_size());
-            },
-            R"delim(
-        Return the genetic values for ancient samples as a readonly 2d 
-        numpy.ndarray.
-        
-        Rows are individuals.  Columns are genetic values.
-        
-        ..  versionadded 0.3.0
-        )delim");
+            });
 }
