+++ fwdpy11/src/ts/simplify.cc
@@ -15,13 +15,11 @@ simplify(const fwdpy11::Population& pop,
 {
     if (pop.tables->genome_length() == std::numeric_limits<double>::max())
         {
-            throw std::invalid_argument(
-                "population is not using tree sequences");
+            throw std::invalid_argument("population is not using tree sequences");
         }
     if (pop.tables->num_nodes() == 0)
         {
-            throw std::invalid_argument(
-                "population has empty TableCollection");
+            throw std::invalid_argument("population has empty TableCollection");
         }
     if (samples.empty())
         {
@@ -30,8 +28,7 @@ simplify(const fwdpy11::Population& pop,
     if (std::any_of(samples.begin(), samples.end(),
                     [&pop](const fwdpp::ts::table_index_t s) {
                         return s == fwdpp::ts::NULL_INDEX
-                               || static_cast<std::size_t>(s)
-                                      >= pop.tables->num_nodes();
+                               || static_cast<std::size_t>(s) >= pop.tables->num_nodes();
                     }))
         {
             throw std::invalid_argument("invalid sample list");
@@ -40,76 +37,26 @@ simplify(const fwdpy11::Population& pop,
     fwdpp::ts::table_simplifier<fwdpp::ts::std_table_collection> simplifier{};
     auto rv = simplifier.simplify(t, samples);
     t.build_indexes();
-    return py::make_tuple(std::move(t), fwdpy11::make_1d_array_with_capsule(
-                                            std::move(rv.first)));
+    return py::make_tuple(std::move(t),
+                          fwdpy11::make_1d_array_with_capsule(std::move(rv.first)));
 }
 
 void
 init_simplify_functions(py::module& m)
 {
-    m.def("simplify", &simplify, py::arg("pop"), py::arg("samples"),
-          R"delim(
-            Simplify a TableCollection stored in a Population.
-
-            :param pop: A :class:`fwdpy11.PopulationBase`
-            :param samples: A list of samples (node indexes).
-                
-            :return: The simplified tables and array mapping input sample IDs to output IDS
-
-            :rtype: tuple
-
-            Note that the samples argument is agnostic with respect to the time of
-            the nodes in the input tables. Thus, you may do things like simplify
-            to a set of "currently-alive" nodes plus some or all ancient samples by
-            including some node IDs from :attr:`fwdpy11.DiploidPopulation.ancient_sample_metadata`.
-            
-            If the input contains ancient samples, and you wish to include them in the output,
-            then you need to include their IDs in the samples argument.
-
-            .. note::
-
-                Due to node ID remapping, the metadata corresponding to nodes becomes a bit more
-                difficult to look up.  You need to use the output ID map, the original IDs, and 
-                the population's metadata containers.
-
-            .. deprecated:: 0.3.0
-
-                Prefer :func:`fwdpp.simplify_tables`
-
-            
-            .. versionchanged:: 0.3.0
-
-                Ancient samples are no longer kept by default
-
-            .. versionchanged:: 0.5.0
-
-                No longer requires a :class:`MutationVector` argument.
-
-            )delim");
+    m.def("_simplify", &simplify, py::arg("pop"), py::arg("samples"));
 
     m.def(
-        "simplify_tables",
+        "_simplify_tables",
         [](const fwdpp::ts::std_table_collection& tables,
            const std::vector<fwdpp::ts::table_index_t>& samples) -> py::tuple {
             auto t(tables);
             fwdpp::ts::table_simplifier<fwdpp::ts::std_table_collection> simplifier{};
             auto rv = simplifier.simplify(t, samples);
             t.build_indexes();
-            return py::make_tuple(std::move(t),fwdpy11::make_1d_array_with_capsule(std::move(rv.first)));
+            return py::make_tuple(
+                std::move(t), fwdpy11::make_1d_array_with_capsule(std::move(rv.first)));
         },
-        py::arg("tables"), py::arg("samples"),
-        R"delim(
-          Simplify a TableCollection.
-          
-          :param pop: A table collection.
-          :type pop: :class:`fwdpy11.TableCollection`
-          :param samples: list of samples
-          :type list: list-like or array-like
-
-          :returns: A simplified TableCollection and an array containing remapped sample ids.
-          :rtype: tuple
-
-          .. versionadded:: 0.3.0
-          )delim");
+        py::arg("tables"), py::arg("samples"));
 }
